Module tidyup
